After modifying the code including the nifty8 changes and the eg model changes, we temporarily switched back to nifty7 in order to run some tests.

+++ src/model_library/noise_models/simple_variable_noise.py
@@ -1,4 +1,5 @@
-import nifty8 as ift
+import nifty7 as ift
+
 import numpy as np
 
 from ..Model import Model
@@ -8,7 +9,7 @@ class SimpleVariableNoise(Model):
 
     def __init__(self, target_domain, noise_cov, alpha, q, name=''):
         if isinstance(noise_cov, ift.Field):
-            self.noise_cov = ift.makeOp(noise_cov,sampling_dtype=float)
+            self.noise_cov = ift.makeOp(noise_cov)
         else:
             self.noise_cov = ift.makeOp(ift.Field(self.target_domain, noise_cov),sampling_dtype=float)
         self.alpha = alpha
