+++ src/operators/InverseGamma.py
@@ -1,5 +1,5 @@
 import numpy as np
-import nifty8 as ift
+import nifty7 as ift
 from scipy.stats import invgamma, norm
 from scipy.special import gammaincinv, gammainccinv
 
@@ -90,7 +90,7 @@ class InverseGammaOperator(ift.Operator):
                 # Derivative of linear interpolation
                 der = self._deriv[fi] * res
 
-                jac = ift.makeOp(ift.Field(self._domain, der),sampling_dtype=float)
+                jac = ift.makeOp(ift.Field(self._domain, der))
                 jac = jac(x.jac)
                 return x.new(points, jac)
